[5.3] Actionlogs User filter
Loading all the users how have ever made an action in a gigantic list is non-performant on a site with a large number of users.

This PR changes the field to a user select. The downside is that you could select a user who has not made any actions but that is outweighed by the performace gain

Signed-off-by: BrianTeeman <brian@teeman.net>