[Cypress] Refactoring SEF Test to Avoid Using the Joomla Command-Line Client Tool (#44656)
* [Cypress] Remove Joomla command line dependency

With the `SefPlugin.cy.js` test specification, the Joomla command line client tool was added to the Joomla system tests for the first time. This is an additional complexity and dependency. Especially if the test environment is containerised and Cypress container has no PHP installed.  And PHP version needs to be 8.1 and additional modules like `php-simplexml` are needed. And with the Docker access rights, `configuration.php` must be opened for writing (`chmod 644`).
And then there is the locally running Cypress GUI to be configured...

Therefore switched from using the Joomla command line client tool to using the existing `config_setParameter` Cypress command. The original creator of [44253](https://github.com/joomla/joomla-cms/pull/44253) could not know as
it is not documented (will be solved with another PR and together with enabling `mod_rewrite` and the `AllowOverride All` needed configuration).

Since sometimes the first SEF test failed when the entire Joomla System Tests was performed, the following was implemented:
* The operations of the `writeRelativeFile` Cypress task are wrapped in a Promise to ensure that they are only resolved when everything is done
* The `config_setParameter` Cypress custom command to return a Cypress chainable
* Chaining has been implemented everywhere in `SefPlugin.cy.js`

For this test spec the Cypress best practice to use `beforeEach` to ensure tests is added to
run independently from one another (and needed for other places in the Joomla System Tests – one more PR).

As a small side effect the SEF test spec with 9 tests runs faster, e.g. 3 instead of 18 seconds before.

Tested with own target Joomla 5.2-dev three times the single test spec on:
* Intel macOS 15.2 using Apache/MariaDB
* Apple Silicon macOS 15.2 using Apache/MariaDB
* Windows 11 24H2 using Laragon/Apache/MySQL
```
npx cypress run --spec tests/System/integration/plugins/system/sef/SefPlugin.cy.js
```
And once the overall test suite:
```
npm run cypress:run
```

* writeRelativeFile() header adapted return Promise

Thank @heekc29 for the hint

* Delete cy.log in custom command

---------

Co-authored-by: Allon Moritz <allon.moritz@digital-peak.com>