[5.3] Logical css joomla tabs (#44607)
* [5.3] Logical CSS properties

Use logical CSS properties to avoid having to set different LTR and RTL css

This is a redo of #36034 which could not be merged at the time as browsers didnt support logical css properties for floats. They do now https://caniuse.com/?search=float-inline

The css changed is for the highlighted areas of the screenshot. There is no visual change

As this is a scss change then you will need to either rebuild the css or test with a prebuilt package

Signed-off-by: BrianTeeman <brian@teeman.net>

* position

Signed-off-by: BrianTeeman <brian@teeman.net>

* stylelint

Signed-off-by: BrianTeeman <brian@teeman.net>

---------

Signed-off-by: BrianTeeman <brian@teeman.net>